refactor(web-resources): separate bootstrap from controller
